Lower Roslyn floor and add host-pinned package smoke tests (#135)
* Update RelaxVersioner and remove CI single-process build

* Refactor analyzer evaluation pipeline (#133)

* Add packed package smoke test (#132)

* Close ADR-013 traceability (#130)

* Clarify future design space in specs (#131)

* Lower Roslyn floor and add host-pinned package smoke tests

* Handle first-run dotnet version banner in smoke test

* Fix packed package smoke test exit code handling

* Address PR review feedback
