+++ src/DependencyContractAnalyzer/Helpers/DependencyCollector.cs
@@ -267,28 +267,3 @@ internal static class DependencyCollector
         dependencies.Add(new DependencyDescriptor(namedType, dependencyKind));
     }
 }
-
-internal readonly struct DependencyDescriptor
-{
-    public DependencyDescriptor(INamedTypeSymbol type, DependencyKind kind)
-    {
-        Type = type;
-        Kind = kind;
-    }
-
-    public INamedTypeSymbol Type { get; }
-
-    public DependencyKind Kind { get; }
-}
-
-internal enum DependencyKind
-{
-    ConstructorParameter,
-    MethodParameter,
-    Property,
-    Field,
-    ObjectCreation,
-    StaticMemberAccess,
-    BaseType,
-    InterfaceImplementation,
-}
